+++ SADPPBV/src/main/java/com/belarek/SADPPBV/repository/PontoRepository.java
@@ -3,5 +3,7 @@ package com.belarek.SADPPBV.repository;
 import com.belarek.SADPPBV.entity.Ponto;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PontoRepository extends JpaRepository<Ponto, Long> {
+import java.util.Optional;
+
+public interface PontoRepository extends JpaRepository<Ponto, Integer> {
 }
